backend/requirements.txt
# ai
langchain==0.0.337
openai==1.3.3
streamlit==1.28.1
transformers==4.35.2
torch==2.1.1
# houesekeeping
python-dotenv==1.0.0
Flask-Cors==4.0.0
# vector database
pinecone-client==2.2.4
# testing
pytest==7.1.3